+++ ChangeLog.md
@@ -2,6 +2,7 @@ v1.0.1
 ======
 
 + Fix issue where non-class files are dropped when using relocation ([Issue #58](https://github.com/johnrengelman/shadow/issues/58))
++ Do not create a / directory inside the output jar.
 
 v1.0.0
 ======