+++ ChangeLog.md
@@ -3,6 +3,7 @@ v1.0.1
 
 + Fix issue where non-class files are dropped when using relocation ([Issue #58](https://github.com/johnrengelman/shadow/issues/58))
 + Do not create a / directory inside the output jar.
++ Fix `runShadow` task to evaluate the `shadowJar.archiveFile` property at execution time. ([Issue #60](https://github.com/johnrengelman/shadow/issues/60))
 
 v1.0.0
 ======